runtime: fix dropped preemption requests
There are three fields of the m that can disable preemption. In
newstack, we check all three and unpoison the stack bound if
preemption is disabled. However, this means that whenever we may
transition from preemption being disabled to preemption being enabled,
it's necessary to check the preempt flag and re-poison the stack
bound.

There were a few places where we weren't doing this, so preemption
requests could effectively be dropped. Fix these places and document
the requirement.

I tried to remove m.preemptoff to simplify these conditions, but
m.preemptoff is important because there are cases where we need to
disable preemption, but not appear to be holding locks (e.g., when
calling stopTheWorld).

Change-Id: Ic15f6b5d59fafa6a72e6996e88014b23d703d9e8