./github: create new issue template for gopls
gopls issues typically require additional information to diagnose, and
it seems easiest to ask for that information upfront. As explained on
the Github blog, it seems that one can have multiple templates that can be
linked to, while retaining the default template in ISSUE_TEMPLATE.md, which
is what I've done here.

Source: https://github.blog/2018-01-25-multiple-issue-and-pull-request-templates/

Change-Id: Id45e75c18ddbc4b4ac67a4fa343b5cf13c78f9c4