Update patch set 3
Patch Set 3:

(2 comments)

Josh, I think I'm going to leave the tests as is, but I'll benchmark the two forms just for fun ... the compiler in my head said this version would be faster. :-) My thinking was it only has to create the slice if it needs to (non-nil case).

I slightly prefer the start==stop version over the version which mutates the slice variable, so if benchmarking is a wash I'll leave as is.

Patch-set: 3
Reviewer: Gerrit User 28469 <28469@62eb7196-b449-3ce5-99f1-c037f21e1705>