runtime: initial support for timers in P's
This is preliminary support for storing timers in P's. Each P will
have a slice of timer structs to run. When a P is looking for
something to do, it will run any timers that are ready. When a P has
nothing to do, but does have at least one timer, it will sleep until
that timer is ready. The scheduler routines will wake up a P sleeping
on a timer if there is something for it to do.

The effect of this is to speed up the time it takes to add a new timer
as no real locks are required, just a couple of atomic operations.
It also speeds up the time it takes to run a timer, as no goroutine
communication or scheduling is required.

This CL just adds some initial support for scheduling P's with timers.
No timers are actually added to P's yet. That will come in future CL's.
Also, more scheduling changes follow to address more esoteric situations.

Updates #27707

Change-Id: I3fbfaa3ad583d34d4389fa2ee6fefb66275f6fa2