runtime: make preemption requests robust
Our current preemption mechanism contains races that can clobber
preemption requests, making it necessary to retry preemption requests
until they get through. It involves two main fields: the preempt flags
and the stack guard. In order to trigger a preemption, we set some
preempt flag and then poison the stack guard. When running, goroutines
only check the stack guard (not the preempt flags). But there are many
places where we have to un-poison the stack guard, such as after we
handle a preemption, or when returning from a syscall, or after a
stack copy. Right now, if this races with another thread requesting a
preemption, it can leave the goroutine in a state where preempt flags
are set, but the stack guard is not poisoned.

Currently we deal with this by making STW and stack scan preemptions
retry until the goroutine acknowledges the preemption. These retry
loops aren't great. But worse, we're about to make stack shrinking use
its own preemption request, and that we want to be a fire-and-forget
request. The current mechanism could simply lose this request until
another preemption happens to poison the stack guard.

Hence, this CL reworks how we manipulate preempt flags and the stack
guard to fix this clobbering and make preemption requests robust. It
abstracts these manipulations into three operations: setting a preempt
flag, clearing a preempt flag, and resetting the stack bound. These
operations use careful ordering to make sure preemption requests
aren't clobbered.

As of this CL, preemptions can still be arbitrarily delayed because we
don't drain preemption tasks everywhere we need to. As a result, the
scang retry loop is still necessary to make stack scanning prompt.
We'll fix this shortly.

For #10958, #24543.

Change-Id: If6cff5ba94053687243bd0b6186ca5dfb903643c