runtime: add new adjusttimers function
The adjusttimers function is where we check the adjustTimers field in
the P struct to see if we need to resort the heap. We walk forward in
the heap and find and resort timers that have been modified, until we
find all the timers that were modified to run earlier. Along the way
we remove deleted timers.

This CL also adds the scheduler code that calls adjustTimers.

Updates #27707

Change-Id: I1ef19e7870f1a6a772560edefb4c661f275ba944