cmd/compile: represent slices as (ptr, len, cap-len) in SSA
DO NOT SUBMIT, work in progress.

TODOs:
  * benchmarks
  * update prove and tests
  * more comments
  * more optimizations
  * guard cap bounds checks with len checks to avoid
    loading/calculating cap in hot path

This CL changes the way slices are represented in SSA. Rather than
having a capacity (cap), they now have the field extra (ext), which
is the difference between the capacity and the length.

The advantage of representing slices this way is that pointer-only
reslices now only require updating the pointer and length. This can
help reduce register pressure when doing pointer-only reslicing in
a loop.

For example, the common pattern below now only requires pointer and
length values in the likely path of the loop. The (extra) capacity
is now loop invariant.

    for len(x) > 0 {
        ...
        x = x[n:]
    }

This approach does have disadvantages. Firstly it moves the SSA
representation away from the in-memory representation of slices
which could cause confusion and therefore bugs. Secondly it can
lead to extra arithmetic being required when reslicing the length.

Change-Id: I4c591a95b01c7e6acab4d2d37c021b6566f98c11