runtime: make preemption requests prompt
Now that preemption requests are robust, this CL makes them also
prompt by checking for and processing preemption tasks at all
descheduling points. This will finally enable "fire and forget"
preemption requests, where the runtime can queue a preemption task and
assume that it will execute soon. This is also a step toward
asynchronous preemptions.

This adds one more check to the syscall entry path. However, we're
going to remove the runSafePointFn check in a later CL, so there will
be no net change (the new check is probably slightly faster, too).

This also tweaks scanstack to accept the _Gdead state in addition to
_Gscandead (and do nothing), since we need to drain preemptions on
goroutine exit to maintain bookkeeping.

For #10958, #24543.

Change-Id: I561775534b73a6f4546f2a663a25e05d0deea344