Update patch set 3
Patch Set 3: Run-TryBot+1

> I assume this passes toolstash-check.

Not quite, but the differences are very minor. They are caused by (a) no longer unconditionally creating a zero value and (b) adding an OpCopy to convert from a pointer to an array to a pointer to the element type.

The latest CL rearranges the masking code a bit to keep the order the same as the original code. It also removes the type t from the function signature and gets it from v instead. I think this makes it clearer that t is the type of the value being sliced rather than some other type (e.g. destination type).

Patch-set: 3
Reviewer: Gerrit User 11990 <11990@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Run-TryBot=+1