Update patch set 7
Patch Set 7:

(1 comment)

Could you do a pass over all the types, struct fields, and methods and add comments where it's not super obvious and would help a new reader starting to learn this codebase?

It's easier to do it while it's still in at least one person's head.

Patch-set: 7