Update patch set 7
Patch Set 7:

(1 comment)

> Could you do a pass over all the types, struct fields, and methods and add comments where it's not super obvious and would help a new reader starting to learn this codebase?
> 
> It's easier to do it while it's still in at least one person's head.

Yeah, I've made a few passes and tried to add comments where I thought was necessary. It's a little difficult because I've been looking at the code for so long I'm not feeling confident about which code is intuitive or not to others.

Maybe it'd be worthwhile to meet with someone to do a once-over explaining the code and see what questions come up?

Patch-set: 7
Reviewer: Gerrit User 5440 <5440@62eb7196-b449-3ce5-99f1-c037f21e1705>