Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> (3 comments)

My recollection, meet vs join, is that (at least in the 1980s) you approached your analysis lattice differently depending on which side of the Atlantic Ocean you were -- in the US, people in the Fortran-ish school of optimization started at Top (that is the most optimistic place) and did meets (GLBs, infimums) towards bottom, which requires the most pessimistic treatment.  In England and Europe, analysis/optimism started at Bottom, and proceeded with joins in the pessimistic direction towards Top.  

And I think that the EU convention has gradually taken over, which is fine with me, and thus I believe that this is a join operation since it makes the most pessimistic (most escape-y) interpretation of its inputs.

At least, this is my recollection.

Patch-set: 7
Reviewer: Gerrit User 7061 <7061@62eb7196-b449-3ce5-99f1-c037f21e1705>