Update patch set 1
Patch Set 1:

Thank you for this change Tomuta! Could you please file an issue on the Go issue tracker at https://github.com/golang/go/issues? It is useful for a discussion such as the one
that's going on, on this CL yet that discussion isn't as accessible nor indexed nor 
searchable.
Without a corresponding issue, it might be difficult to later on to examine security issues, regressions and reasons why we made certain decisions.

Thank you.

Patch-set: 1
CC: Gerrit User 5137 <5137@62eb7196-b449-3ce5-99f1-c037f21e1705>