Update patch set 1
Patch Set 1:

> In any case, another reasoning is this: if there already is BOOT_GO_GCFLAGS, why shouldn't there be BOOT_GO_LDFLAGS as well?

Thanks for the detailed explanation of why BOOT_GO_LDFLAGS is useful.  In answer to this question, every knob is something that has to be documented and maintained and fixed when surrounding code changes.  In other words, every knob has a cost.  The question is never "why not add a knob?"  It's always "why add a knob?"

Patch-set: 1