cmd/compile: handle sign/zero extensions in prove, via update method
DO NOT SUBMIT. Editing commit msg to explain/reflect changes.

Array accesses with index types smaller than the machine word size may
involve a sign or zero extension of the index value before bounds
checking. Currently, this defeats prove because the facts about the
original index value don't flow through the sign/zero extension.

Fix this by looking back through value-preserving sign/zero extensions
when adding facts via Update and, where appropriate, applying the same
facts using the pre-extension value. Facts learned in domains which are
not matched to the pre-extension value's type may not be valid when applied
to that value. E.g. When a uint8 is zero-extended to int for indexing,
signed domain facts learned about the int cannot always be applied to the
uint8, despite the fact that uint8 -> int is a value-preserving zeroExt.

Also note that we are unable to distinguish between extensions added by
the compiler for indexing operations, extensions added by the compiler for
implicit conversions, and explicit extensions from the source. This means
we need to consider a wider set of possible combinations of extensions
and types, and determine which facts are appropriate to apply to
pre-extensions values.

This is an alternative to CL 122695 and CL 174309

Fixes #26292. Fixes #29964.

Removes 238 bounds checks from std/cmd.

Change-Id: I1f87c32ee672bfb8be397b27eab7a4c2f304893f