Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> I don't understand what new guarantees this code provides.
> 
> In particular, whether we return EINVAL or not is now data-dependent, which seems wrong. There's no way for the user to decide ahead of time what buffer sizes would be ok, unless they already knew what was in the directory.
> 
> Do we need to require a larger buffer to Getdirentries? Maybe then MINBUFSIZ or whatever constant is exposed to users needs to be increased?  The fact that syscall doesn't promise compatibility makes that possible, but it probably isn't nice if we can avoid it.
> 
> The fact that you're changing the test to make it work means we'll be breaking user code. (Well, user code that's probably broken now anyway, but maybe not so blatantly.) Can we find a way to do it without breaking users' code?

> The fact that you're changing the test to make it work means we'll be breaking user code. (Well, user code that's probably broken now anyway, but maybe not so blatantly.) Can we find a way to do it without breaking users' code?

I've changed the test so it wouldn't have to handle EINVAL which is a valid result for an insufficient buffer.

We wrap the real syscall with a shim that may return EINVAL when the buffer is not large enough to hold the result. Similarly to how the underlying syscall would return EINVAL if passed a bad fd or a small buffer.

I don't see how I'm breaking user's code..

Patch-set: 5
Reviewer: Gerrit User 5781 <5781@62eb7196-b449-3ce5-99f1-c037f21e1705>