Update patch set 5
Patch Set 5:

> If I'm reading it correctly, the manpage says any size MINBUFSIZ or larger is ok.
> I'm sure many callers pass a MINBUFSIZ-sized buffer and have no ability to size the buffer larger on an EINVAL error code.

MINBUFSIZ == DIRBLKSIZ ?

We don't expose it to the user, the libc of most BSD derivatives has a header with:

/usr/include/dirent.h
  /* definitions for library routines operating on directories. */
  #define	DIRBLKSIZ	1024 

It derives from the UFS filesystem I think, the modern equivalent would be to use the statvfs.f_frsize value.

https://www.freebsd.org/cgi/man.cgi?query=getdirentries doesn't mention DIRBLKSIZ either:

> The nbytes argument must be greater than or equal to
> the block size associated with the file, see stat(2). Some file
> systems may not support these system calls with buffers smaller than this
> size.

EINVAL is a valid result, precisely as in the netbsd builder case which now no longer fails because the buffer is large enough.

Patch-set: 5
Reviewer: Gerrit User 5781 <5781@62eb7196-b449-3ce5-99f1-c037f21e1705>