Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Although the time it prints is in some sense confusing, I found this sometimes useful. I hope we don't delete it before we have a better alternative.
> 
> When/how is it useful? (And what is useful? The existing prints, or the timing routine?) I'm happy to discuss an alternative, but I'm not sure what we are trying to replace or why. :)
> 
> Maybe just using the log package, which adds timestamps, would suffice?

The time spent in each pass of the linker. It is confusing in that many passes print cputime at the beginning, but not all. Using the log package would probably not be a good replacement, as the absolute timestamps are not that easy to read. I think it may be a good idea to decouple the time prints from -v, though, if that was the concern. (maybe with a different flag)

Patch-set: 2
Reviewer: Gerrit User 13315 <13315@62eb7196-b449-3ce5-99f1-c037f21e1705>