cmd/link: use prefix/suffix aware methods more often
DO NOT SUBMIT
DO NOT REVIEW

Add new method NameEqString for checking to see if a given symbol name
is equal to a string (to reduce allocation); introduce calls to this
method in various places so as to avoid having to allocate a string so
as to compare it to a fixed symname. Switch over HasPrefix/HasSuffix
sym name methods to use faster versions.

Change-Id: Iaf99d30a12c3b02ebbe8389402bd2164a0098c6e