Update patch set 9
Patch Set 9:

I think this is ready to be reviewed again. Benchmarking now streams, too. I decided to resolve https://github.com/golang/go/issues/24929#issuecomment-436305273 by printing the name before the test run, then streaming happens, then print the name again + results. Let me know if y'all think this is the wrong approach.

It has a lot less tests than I'd like, but after a couple of hours struggling to make common.writer and common.output work the way I expect, I'm throwing the towel in for the moment. What I want is a writer that writes to a buffer (output), which gets flushed at end of test. In chatty, I would swap out to a writer that writes to stdout and no buffer. This way I can stick in a test writer and capture all the good output that currently the tests are unable to introspect.

However, in the current impl, in some places writer writes to output, and in other places we stick output in the writer. I'm sure there's a good reason, it's just a bit confusing for a first timer in this part of the codebase. I'll take another look later this week and see if I can understand it better. :)

For now, though, would appreciate any guidance or review on the general direction this is going.

Patch-set: 9
Reviewer: Gerrit User 26615 <26615@62eb7196-b449-3ce5-99f1-c037f21e1705>