Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > If I'm reading it correctly, the manpage says any size MINBUFSIZ or larger is ok.
> > I'm sure many callers pass a MINBUFSIZ-sized buffer and have no ability to size the buffer larger on an EINVAL error code.
> 
> MINBUFSIZ == DIRBLKSIZ ?
> 

Right, sorry.

> We don't expose it to the user, the libc of most BSD derivatives has a header with:
> 

Well, it is exposed to the user as syscall.DIRBLKSIZ in Go.

> /usr/include/dirent.h
>   /* definitions for library routines operating on directories. */
>   #define	DIRBLKSIZ	1024 
> 
> It derives from the UFS filesystem I think, the modern equivalent would be to use the statvfs.f_frsize value.
> 
> https://www.freebsd.org/cgi/man.cgi?query=getdirentries doesn't mention DIRBLKSIZ either:
> 
> > The nbytes argument must be greater than or equal to
> > the block size associated with the file, see stat(2). Some file
> > systems may not support these system calls with buffers smaller than this
> > size.

So what is the block size associated with a directory? i.e., what does stat return, and will this change cause us to return EINVAL for that size? Why or why not?

> 
> EINVAL is a valid result, precisely as in the netbsd builder case which now no longer fails because the buffer is large enough.

There are two separate issues here:
1. Are we allowed to return EINVAL in this case?
2. Will it break existing programs?

I'm more worried about the second point. syscall is exempt from the Go1 guarantee, so we are allowed to break users' existing programs.  But we don't want to do that if we don't have to. It would be illustrative to find actual *bsd programs that call getdirentries, but my suspicion is that they allocate buffers of size 1024 or sycsall.DIRBLKSIZ and have no ability to size the buffer larger and try again if EINVAL is returned.

I'm not yet convinced that we're ok on the first point either.

Patch-set: 5
Reviewer: Gerrit User 5200 <5200@62eb7196-b449-3ce5-99f1-c037f21e1705>