cmd/go/internal/base: add a Logf function and funnel stderr output through it
GOAUTH will potentially add user-interactive prompts while a 'go'
command is running. If we log to stderr at the same time, that will
disrupt and potentially confuse the user.

Instead, we can guard the user interaction with a mutex: that way, we
can be sure we won't stomp on the user's terminal while they're
entering credentials.

Updates #26232

Change-Id: I2eb099afbfe68479f3d3b9dde0611b99b241e23f