go/internal/gccgoimporter: improve alias handling for anonymous fields
The code in gccgoimporter that deals with anonymous structure fields
record the fact that a field is anonymous, but tries to install a proxy
name for the field based on the name of the type used to declare
the field. If that type was an alias, however, don't try to recover the
name.

Change-Id: I9b7369ed558a288b56d85170c6f1144daf5228eb